Reorganise pyswervedrive to have non-repetitive name
