+++ autonomous/autonomous.py
@@ -10,7 +10,7 @@ from automations.cube import CubeManager
 from automations.motion import ChassisMotion
 from components.intake import Intake
 from components.vision import Vision
-from pyswervedrive.swervechassis import SwerveChassis
+from pyswervedrive.chassis import Chassis
 from utilities.imu import IMU
 
 from wpilib import SmartDashboard
@@ -26,14 +26,14 @@ class OverallBase(AutonomousStateMachine):
     vision: Vision
     intake: Intake
     imu: IMU
-    chassis: SwerveChassis
+    chassis: Chassis
     ds: wpilib.DriverStation
 
     # automations
     motion: ChassisMotion
     cubeman: CubeManager
 
-    START_Y_COORDINATE = 3.4 - SwerveChassis.WIDTH/2
+    START_Y_COORDINATE = 3.4 - Chassis.WIDTH/2
 
     # Coordinates of various objectives no the field
     # Default to those for LEFT HAND SIDE of the field
@@ -43,7 +43,7 @@ class OverallBase(AutonomousStateMachine):
     # CUBE_PICKUP_1 = (6.2, 1.7)
     CUBE_PICKUP_1 = (6.2, 0.9)
     CUBE_PICKUP_2 = (6.2, 1.03)
-    SWITCH_DEPOSIT = (5 + SwerveChassis.LENGTH/2, 1.2)
+    SWITCH_DEPOSIT = (5 + Chassis.LENGTH/2, 1.2)
     SCALE_INIT_WAYPOINT = (6, 3)
 
     SWITCH_DEPOSIT_ORIENTATION = -math.pi
@@ -54,16 +54,16 @@ class OverallBase(AutonomousStateMachine):
     PICKUP_WAYPOINT_X = 6.5
     CROSS_POINT = (6.1, START_Y_COORDINATE)
     OPP_CROSS_POINT = (6.1, -START_Y_COORDINATE)
-    DRIVE_BY_SWITCH_POINT = (4, 2 + SwerveChassis.LENGTH/2)
+    DRIVE_BY_SWITCH_POINT = (4, 2 + Chassis.LENGTH/2)
     """
     START_Y_COORDINATE = 1
-    SCALE_DEPOSIT = (6 - SwerveChassis.LENGTH/2, 1)
-    # CUBE_PICKUP_1 = (3 + SwerveChassis.LENGTH/2, 0.5)
-    # CUBE_PICKUP_2 = (3 + SwerveChassis.LENGTH/2, -0.5)
+    SCALE_DEPOSIT = (6 - Chassis.LENGTH/2, 1)
+    # CUBE_PICKUP_1 = (3 + Chassis.LENGTH/2, 0.5)
+    # CUBE_PICKUP_2 = (3 + Chassis.LENGTH/2, -0.5)
     # CUBE_PICKUP_1 = (3+1, 0.5)
     CUBE_PICKUP_1 = (3+1, 0.5)
     CUBE_PICKUP_2 = (3+1, -0.5)
-    SWITCH_DEPOSIT = (3 + SwerveChassis.LENGTH/2, 0)
+    SWITCH_DEPOSIT = (3 + Chassis.LENGTH/2, 0)
 
     SWITCH_DEPOSIT_ORIENTATION = -math.pi
 
@@ -72,7 +72,7 @@ class OverallBase(AutonomousStateMachine):
     PICKUP_WAYPOINT_X = 5
     SWITCH_TO_CUBE_POINT = (PICKUP_WAYPOINT_X, 0.8)
 
-    DRIVE_BY_SWITCH_POINT = (2, 0.5 + SwerveChassis.LENGTH/2)
+    DRIVE_BY_SWITCH_POINT = (2, 0.5 + Chassis.LENGTH/2)
     CROSS_POINT = (4.5, 1)
     OPP_CROSS_POINT = (4.5, -1)
     """
@@ -132,7 +132,7 @@ class OverallBase(AutonomousStateMachine):
         self.fms_switch = game_data[0]  # L or R
         self.init_plate_locations()
 
-        self.chassis.odometry_x = SwerveChassis.LENGTH / 2
+        self.chassis.odometry_x = Chassis.LENGTH / 2
 
         self.cube_number = 0
